chore: remove old run command
